style(bridge): move monitors to their directory
