+++ bridge/src/index.ts
@@ -4,13 +4,13 @@ import { init } from "@sentry/node";
 import { BurnEventResult } from "./types/burn-event-result";
 import { IWrappedNCGMinter } from "./interfaces/wrapped-ncg-minter";
 import { INCGTransfer } from "./interfaces/ncg-transfer";
-import { EthereumBurnEventMonitor } from "./ethereum-burn-event-monitor";
+import { EthereumBurnEventMonitor } from "./monitors/ethereum-burn-event-monitor";
+import { NineChroniclesTransferredEventMonitor } from "./monitors/nine-chronicles-transferred-event-monitor";
 import { NCGTransfer } from "./ncg-transfer";
 import { WrappedNCGMinter } from "./wrapped-ncg-minter";
 import { wNCGTokenAbi } from "./wrapped-ncg-token";
 import HDWalletProvider from "@truffle/hdwallet-provider";
 import { HeadlessGraphQLClient } from "./headless-graphql-client";
-import { NineChroniclesTransferredEventMonitor } from "./nine-chronicles-transferred-event-monitor";
 import { BlockHash } from "./types/block-hash";
 import { TxId } from "./types/txid";
 import { IHeadlessHTTPClient } from "./interfaces/headless-http-client";
